update2024-04-19-16:48

+++ resources/js/pages/Store.vue
@@ -23,14 +23,20 @@
           <button type="button" class="btn btn-secondary" @click="CancelStore()">ຍົກເລິກ</button>
       </div>
       <div class="row">
-        <div class="col-md-4">Image</div>
+            <div class="col-md-4 text-center" style=" position: relative; ">
+                <button type="button" class="btn rounded-pill btn-icon btn-danger img-remove" @click="RemoveImage()"  v-if="FormStore.image" >
+                    <i class='bx bxs-x-circle fs-3'></i>
+              </button>
+               <img :src="image_pre" @click="$refs.img_store.click()" class=" rounded cursor-pointer " style=" width: 80%;" alt="">
+               <input type="file" ref="img_store" style=" display: none;" @change="onSelect" >
+            </div>
         <div class="col-md-8"> {{FormStore}}
           <h1></h1>
             <label class="form-label fs-6">ຊື່ສິນຄ້າ: <span class="text-danger">*</span></label>
             <input type="text" class="form-control mb-2" v-model="FormStore.name" placeholder=".....">
 
             <label class="form-label fs-6">ຈຳນວນ: <span class="text-danger">*</span></label>
-            <input type="text" class="form-control mb-2" v-model="FormStore.amount" placeholder=".....">
+            <input type="text" class="form-control mb-2" v-model="FormStore.amount" placeholder="....." >
 
             <div class="row">
               <div class="col-md-6">
@@ -55,7 +61,7 @@
             <i class='bx bx-sort-up fs-4 me-2' v-if="Sort == 'asc'"></i>
             <i class='bx bx-sort-down fs-4 me-2' v-if="Sort == 'desc'"></i>
             </div>
-            <select id="defaultSelect" class="form-select">
+            <select id="defaultSelect" v-model="list_num" @change="GetStore()" class="form-select">
               <option value="5">5</option>
               <option value="10">10</option>
               <option value="20">20</option>
@@ -63,15 +69,15 @@
             </select>
           </div>
           <div class="d-flex">
-            <input type="text" class="from-control me-2" placeholder="ຄົ້ນຫາ...">
+            <input type="text" class="from-control me-2" placeholder="ຄົ້ນຫາ..." v-model="search" @keyup.enter="GetStore()">
             <button type="button" class=" btn btn-primary" @click="AddStore()"> <i class='bx bx-user-plus' ></i> ເພີ່ມຂໍ້ມູນ </button>
           </div>
       </div>
       <table class="table table-bordered">
         <thead>
           <tr>
-            <th>ID</th>
-            <th>ຮູບພາບ</th>
+            <th width="80">ID</th>
+            <th width="120">ຮູບພາບ</th>
             <th>ຊື່ສິນຄ້າ</th>
             <th>ລາຄາຊື້</th>
             <th>ຈັດການ</th>
@@ -87,13 +93,16 @@
           </tr>
         </tbody>
         <tbody v-else>
-          <tr v-for="list in StoreData" :key="list.id">
+          <tr v-for="list in StoreData.data" :key="list.id">
             <td>{{ list.id }}</td>
-            <td>{{ list.image }}</td>
+            <td>
+              <img :src="url+'/assets/img/'+list.image" v-if="list.image" class=" rounded w-100" alt="">
+              <img :src="url+'/assets/img/no_img.jpg'" v-else class=" rounded w-100" alt="">
+            </td>
             <td>
               {{ list.name }}
             </td>
-            <td>{{ list.price_buy }}</td>
+            <td>{{ formatPrice(list.price_buy) }}</td>
             <td>
               <div class="dropdown">
                 <button type="button" class="btn p-0 dropdown-toggle hide-arrow" data-bs-toggle="dropdown"><i class="bx bx-dots-vertical-rounded"></i></button>
@@ -108,6 +117,8 @@
 
         </tbody>
       </table>
+
+      <pagination :pagination="StoreData" :offset="4" @paginate="GetStore($event)" />
     </div>
   </div>
 </div>
@@ -115,10 +126,14 @@
 </template>
 
 <script>
+
+import Cleave from 'vue-cleave-component';
 import axios from 'axios';
 import { useStore } from '../store/auth';
+import Pagination from '../components/Pagination.vue';
 
 export default {
+  components: { Pagination },
   name: 'Minipos12Store',
   setup(){
     const store = useStore()
@@ -126,12 +141,15 @@ export default {
   },
   data() {
     return {
+      search:'',
+      list_num: 5,
       Sort:"asc",
       StoreData:[],
       EditID:'',
       ShowFrom:false,
       FormStore:{
         name:'',
+        image:'',
         amount:'',
         price_buy:'',
         price_sell:''
@@ -139,7 +157,19 @@ export default {
       FormType: true,
       Loading: false,
       Loading_table: false,
-
+            //   options:{
+            //     //   prefix: '₭ ',
+            //       numeral: true,
+            //       numeralPositiveOnly: true,
+            //       noImmediatePrefix: true,
+            //       rawValueTrimPrefix: true,
+            //       numeralIntegerScale: 10,
+            //       numeralDecimalScale: 2,
+            //       numeralDecimalMark: ',',
+            //       delimiter: '.'
+            // },
+            image_pre: window.location.origin+"/assets/img/add_img.png",
+            url: window.location.origin
     };
   },
 
@@ -162,6 +192,25 @@ export default {
   },
 
   methods: {
+          RemoveImage(){
+
+            this.FormStore.image = ""
+            this.image_pre = window.location.origin+"/assets/img/add_img.png"
+
+            },
+        onSelect(event){
+            // console.log(event)
+            this.FormStore.image = event.target.files[0]
+            let reder = new FileReader()
+            reder.readAsDataURL(this.FormStore.image)
+            reder.addEventListener("load", function(){
+                this.image_pre = reder.result
+            }.bind(this),false)
+        },
+      formatPrice(value) {
+            let val = (value / 1).toFixed(0).replace(",", ".");
+            return val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ".");
+        },
       ChangeSort(){
           if(this.Sort == "asc"){
               this.Sort = "desc"
@@ -175,6 +224,8 @@ export default {
           this.FormStore.amount='',
           this.FormStore.price_buy='',
           this.FormStore.price_sell='',
+          this.image='',
+          this.image_pre=this.url + "/assets/img/add_img.png",
           this.ShowFrom = true,
           this.FormType = true
       },
@@ -184,20 +235,36 @@ export default {
       SaveStore(){
         if(this.FormType){
             //add
-            console.log('Go!!');
+            // console.log('Go!!');
             this.Loading = true
-            setTimeout(() => {
-              console.log('Run...');
+            // setTimeout(() => {
+            //   console.log('Run...');
               axios.post("api/store/add",
-                this.FormStore,{ headers:{ Authorization: 'Bearer '+ this.store.get_token}}
+                this.FormStore,{ headers:{"content-type":"multipart/form-data", Authorization: 'Bearer '+ this.store.get_token}}
                 ).then((res)=>{
                     
                     this.Loading = false
                     if(res.data.success){
                         this.ShowFrom = false
                         this.GetStore()
+                        this.$swal({
+                          toast: true,
+                          position: 'top-end',
+                          icon: 'success',
+                          title: res.data.message,
+                          showConfirmButton: false,
+                          timer: 1000
+                        })
                     }else{
-
+                        this.$swal({
+                          // toast: true,
+                          position: 'center',
+                          icon: 'error',
+                          title: 'ຜິດພາດ',
+                          text: res.data.message,
+                          showConfirmButton: false,
+                          timer: 1000
+                        })
 
                     }
                 }).catch((err)=>{
@@ -211,20 +278,36 @@ export default {
                         this.$router.push("/login")
                     }
                 })
-            }, 2000);
+            // }, 2000);
 
         }else{
               this.Loading = true
               axios.post(`api/store/update/${this.EditID}`,
-                this.FormStore,{ headers:{ Authorization: 'Bearer '+ this.store.get_token}}
+                this.FormStore,{ headers:{"content-type":"multipart/form-data", Authorization: 'Bearer '+ this.store.get_token}}
                 ).then((res)=>{
                     
                     this.Loading = false
                     if(res.data.success){
                         this.ShowFrom = false
                         this.GetStore()
+                        this.$swal({
+                          toast: true,
+                          position: 'top-end',
+                          icon: 'success',
+                          title: res.data.message,
+                          showConfirmButton: false,
+                          timer: 1000
+                        })
                     }else{
-
+                        this.$swal({
+                          // toast: true,
+                          position: 'center',
+                          icon: 'error',
+                          title: 'ຜິດພາດ',
+                          text: res.data.message,
+                          showConfirmButton: false,
+                          timer: 1000
+                        })
                     }
                 }).catch((err)=>{
                     this.Loading = false
@@ -249,6 +332,14 @@ export default {
             this.FormStore = res.data
             this.ShowFrom = true
 
+            //show image form system
+
+            if(res.data.image){
+              this.image_pre = this.url + "/assets/img/" + res.data.image
+            }else{
+              this.image_pre = this.url + "/assets/img/no_img.jpg"
+            }
+
         }).catch((err)=>{
               console.log(err)
               if(err){
@@ -264,35 +355,68 @@ export default {
       },
 
       DeleteStore(id){
-          axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization: 'Bearer '+ this.store.get_token}}).then((res)=>{
+        this.$swal({
+            title: "ທ່ານແນ່ໃຈບໍ່?",
+            text: "ທີ່ຈະລຶບຂໍ້ມູນນີ້!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "ຢືນຢັນ!",
+            cancelButtonText: "ຍົກເລິກ"
+            }).then((result) => {
+              if (result.isConfirmed) {
 
-              if(res.data.success){
-                  this.GetStore()
-              }else{
 
-              }
+                axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization: 'Bearer '+ this.store.get_token}}).then((res)=>{
 
-          }).catch((err)=>{
-              console.log(err)
-              if(err){
-                    if(err.response.status == 401){
-                        this.store.remove_token()
-                        this.store.remove_user()
-                        localStorage.removeItem("web_token")
-                        localStorage.removeItem("web_user")
-                        this.$router.push("/login")
-                      }
+                    if(res.data.success){
+                        this.GetStore()
+                              this.$swal({
+                                toast: true,
+                                position: 'top-end',
+                                icon: 'success',
+                                title: res.data.message,
+                                showConfirmButton: false,
+                                timer: 1000
+                              })
+                    }else{
+                              this.$swal({
+                                // toast: true,
+                                position: 'center',
+                                icon: 'error',
+                                title: 'ຜິດພາດ',
+                                text: res.data.message,
+                                showConfirmButton: false,
+                                timer: 1000
+                              })
                     }
-          })
+
+                }).catch((err)=>{
+                    console.log(err)
+                    if(err){
+                          if(err.response.status == 401){
+                              this.store.remove_token()
+                              this.store.remove_user()
+                              localStorage.removeItem("web_token")
+                              localStorage.removeItem("web_user")
+                              this.$router.push("/login")
+                            }
+                          }
+                    })
+          }
+        })
+
       },
 
-      GetStore(){
+      GetStore(page){
         this.Loading_table = true
-          axios.get(`api/store`,{ headers:{ Authorization: 'Bearer '+ this.store.get_token}}).then((res)=>{
+          axios.get(`api/store?page=${page}&sort=${this.Sort}&list_num=${this.list_num}&search=${this.search}`,{ headers:{ Authorization: 'Bearer '+ this.store.get_token}}).then((res)=>{
               this.Loading_table = false
               this.StoreData = res.data
 
           }).catch((err)=>{
+            this.Loading_table = false
               console.log(err)
               if(err){
                     if(err.response.status == 401){
@@ -309,9 +433,24 @@ export default {
 
   created(){
     this.GetStore()
-  }
+  },
+
+  watch:{
+    search(){
+      if(this.search == ''){
+        this.GetStore()
+      }
+    }
+  },
+    components: {
+      Cleave
+    }
 };
 </script>
-<style lang="scss" scoped>
-
+<style scoped>
+    .img-remove{
+        position: absolute;
+        top: 5px;
+        right: 5px;
+    }
 </style>
